requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
colorama==0.4.6
discord.py==2.2.2
dnspython==2.3.0
filexdb==1.0.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pip==23.0.1
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
setuptools==67.6.1
wheel==0.38.4
yarl==1.8.2
